+++ app/src/androidTest/java/fi/finhua/android/badminton/scoreboard/ApplicationTest.java
@@ -1,4 +1,4 @@
-package fi.fenhua.android.badminton.scoreboard;
+package fi.finhua.android.badminton.scoreboard;
 
 
 import android.app.Application;
