+++ app/src/main/java/fi/finhua/android/badminton/scoreboard/FABBehavior.java
@@ -1,4 +1,4 @@
-package fi.fenhua.android.badminton.scoreboard;
+package fi.finhua.android.badminton.scoreboard;
 
 import android.content.Context;
 import android.support.design.widget.AppBarLayout;
