+++ app/src/main/java/fi/finhua/android/badminton/scoreboard/Game.java
@@ -1,4 +1,4 @@
-package fi.fenhua.android.badminton.scoreboard;
+package fi.finhua.android.badminton.scoreboard;
 
 import java.util.List;
