+++ app/src/main/java/fi/finhua/android/badminton/scoreboard/Utils.java
@@ -1,4 +1,4 @@
-package fi.fenhua.android.badminton.scoreboard;
+package fi.finhua.android.badminton.scoreboard;
 
 import android.content.Context;
 import android.content.res.TypedArray;
