+++ build.gradle
@@ -23,7 +23,7 @@ ext {
     compileSdkVersion = 24
     buildToolsVersion = "24.0.2"
 
-    applicationId = "fi.fenhua.android.badminton.scoreboard"
+    applicationId = "fi.finhua.android.badminton.scoreboard"
     minSdkVersion = 15 // Android 4.0.4
     targetSdkVersion = 24 // Android 7.0
